size_t changed in object id
git-svn-id: file:///home/roberto/Git/svn@7103 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 883eceb27b818abd1f60ee411c263a4b52996a54

+++ trunk/xmipp/libraries/data/image_base.h
@@ -305,12 +305,12 @@ public:
              bool only_apply_shifts = false, MDRow * row = NULL);
 
     /** Read an image from metadata, filename is provided*/
-    int readApplyGeo(const FileName &name, const MetaData &md, long int objId,
+    int readApplyGeo(const FileName &name, const MetaData &md, size_t objId,
              bool readdata=true, int select_img = -1,
              bool only_apply_shifts = false);
 
     /** Read an image from metadata, filename is taken from MDL_IMAGE */
-    int readApplyGeo(const MetaData &md, int objId = -1,
+    int readApplyGeo(const MetaData &md, size_t objId = -1,
              bool readdata=true, int select_img = -1,
              bool only_apply_shifts = false);
